Add local files to project
This commit adds local code for a timekeeping system devised (mostly)
by me. Hxxxxx is the current year from the Holocene epoch (where H1 =
9700BCE); Lx.yy refers to the current lunation, where x refers to the
number of new moons since the spring equinox, and yy refers to the
percentage between the last new moon and the next one; Dxx refers to
the number of days (with the day of the spring equinox being D1);
@xxx.xx is the current .beat time (aka Swatch Internet Time, not made
by me); and the final component is local solar time, where daytime is
represented by Sxx where xx is the percentage of daytime that has
elapsed, and nighttime is Nxx where xx is the percentage of nighttime
that has elapsed.