Fix day count issue again and add conversion func
This commit fixes the day count issue for real, which was actually
caused by the timestamp of the spring equinox being taken as the start
time for every subsequent day. To fix this, we used anchorTime, which
we set to 23:00 UTC the day before the equinox timestamp. It also adds
the conversion function convertGregorianToCustom, which was initially
used to diagnose the cause of the day count issue, and will later be
used for a datetime converter in the UI.